ZORK PART 4 COMMIT
This is the Submission state for Assignment 7 - Zork - Part 4.

This was one hell of an assignment to tackle because I couldn't wrap my head a round of a lot of things in it, but with help from classmates I'm pretty sure I understand what's going on in the code. Me from 2 years ago would be proud.